refactor(checklist): altera path do endpoint de checklist
PL-3461

+++ src/main/java/br/com/zalf/prolog/webservice/v3/frota/checklist/ChecklistDao.java
@@ -35,5 +35,5 @@ public interface ChecklistDao extends JpaRepository<ChecklistEntity, Long> {
                                                     @Nullable final Long codVeiculo,
                                                     boolean incluirRespostas,
                                                     int limit,
-                                                    long offset);
+                                                    int offset);
 }
